Added verification before access iten through id

+++ app/Http/Controllers/ActivityController.php
@@ -25,6 +25,7 @@ class ActivityController extends Controller
 
   public function get(Activity $activity)
   {
+    if($activity->crop->field->farm->client->user->id!=Auth::id()) return response('you dont have access to this activity',400);
     $activity->crop->field->farm->client;
     $activity->activity_type;
     $activity->unity;
@@ -33,6 +34,8 @@ class ActivityController extends Controller
 
   public function update(Activity $activity)
   {
+    if($activity->crop->field->farm->client->user->id!=Auth::id()) return response('you dont have access to this activity',400);
+
     $activity->fill(request()->all());
     $activity->save();
     return $activity;
@@ -40,22 +43,9 @@ class ActivityController extends Controller
 
   public function delete(Activity $activity)
   {
+    if($activity->crop->field->farm->client->user->id!=Auth::id()) return response('you dont have access to this activity',400);
+
     $activity->delete();
     return 'activity deleted';
-  }
-
-  public function activityType(Activity $activity)
-  {
-    return $activity->activity_type;
-  }
-
-  public function crop(Activity $activity)
-  {
-    return $activity->crop;
-  }
-
-  public function unity(Activity $activity)
-  {
-    return $activity->unity;
   }
 }
