+++ app/Http/Controllers/InventoryItenController.php
@@ -26,11 +26,14 @@ class InventoryItenController extends Controller
 
     public function get(InventoryIten $inventoryIten)
     {
+      if($inventoryIten->farm->client->user->id!=Auth::id()) return response('you dont have access to this iten',400);
       return $inventoryIten;
     }
 
     public function update(InventoryIten $inventoryIten)
     {
+      if($inventoryIten->farm->client->user->id!=Auth::id()) return response('you dont have access to this iten',400);
+
       $inventoryIten->fill(request()->all());
       $inventoryIten->save();
       return $farm;
@@ -38,6 +41,8 @@ class InventoryItenController extends Controller
 
     public function delete(InventoryIten $inventoryIten)
     {
+      if($inventoryIten->farm->client->user->id!=Auth::id()) return response('you dont have access to this iten',400);
+      
       $inventoryIten->delete();
       return 'Invetory iten deleted';
     }
